Update environment setup instructions and YAML
The PySyft repository structure has recently been changed;
as a result, installing syft requires a different command now.
Other requirements in the YAML file have been changed
to be consistent with the setup script for syft;
pip was installing different versions of some packages,
uninstalling the versions just installed by conda.
This was not only superfluous,
it also generated an error for one package (flask).